Rename app name to accelerometer
