+++ sensor-graph/accelerometer/build.gradle
@@ -6,13 +6,13 @@ model {
         buildToolsVersion = '23.0.2'
 
         defaultConfig {
-            applicationId    = 'com.android.sensorgraph'
+            applicationId    = 'com.android.accelerometergraph'
             minSdkVersion.apiLevel    = 8
             targetSdkVersion.apiLevel = 23
         }
         ndk {
             platformVersion = 9
-            moduleName = 'sensorgraph'
+            moduleName = 'accelerometergraph'
             toolchain = 'clang'
             stl        = 'gnustl_static'
             // abiFilters.addAll(['armeabi','armeabi-v7a', 'arm64-v8a', 'x86', 'x86-64'])
