+++ sensor-graph/accelerometer/src/main/assets/shader.glslv
