+++ sensor-graph/accelerometer/src/main/res/mipmap-xhdpi/ic_launcher.png
