+++ sensor-graph/accelerometer/src/main/res/mipmap-xxhdpi/ic_launcher.png
