+++ sensor-graph/accelerometer/src/main/res/values/strings.xml
@@ -23,7 +23,7 @@
 
 <resources>
     <!-- Simple strings. -->
-    <string name="sensorgraph_activity">Sensor Graph</string>
+    <string name="accelerometergraph_activity">Accelerometer Graph</string>
 
 </resources>
