+++ sensor-graph/settings.gradle
@@ -1,1 +1,2 @@
-include ':app'
+include 'accelerometer'
+
